+++ public/pages/CreateIndex/components/AliasSelect/AliasSelect.test.tsx
@@ -53,7 +53,6 @@ describe("<AliasSelect /> spec", () => {
         expect(onOptionsChange).toBeCalledWith([
           {
             label: "a",
-            index: "a",
           },
         ]);
         expect(container.firstChild).toMatchSnapshot();
@@ -66,7 +65,9 @@ describe("<AliasSelect /> spec", () => {
 
   it("it should choose options or create one", async () => {
     const { getByTestId } = render(
-      <AliasSelectWithOnchange refreshOptions={() => Promise.resolve({ ok: true, response: [{ alias: "test", query: "test" }] })} />
+      <AliasSelectWithOnchange
+        refreshOptions={() => Promise.resolve({ ok: true, response: [{ alias: "test", index: "123", query: "test" }] })}
+      />
     );
     await waitFor(() => {
       expect(getByTestId("comboBoxInput")).toBeInTheDocument();
@@ -79,18 +80,14 @@ describe("<AliasSelect /> spec", () => {
     await waitFor(() => {
       expect(onChangeMock).toBeCalledTimes(1);
       expect(onChangeMock).toBeCalledWith({
-        test: {
-          query: "test",
-        },
+        test: {},
       });
     });
     await userEvent.type(getByTestId("comboBoxInput"), "test2{enter}");
     await waitFor(() => {
       expect(onChangeMock).toBeCalledTimes(2);
       expect(onChangeMock).toBeCalledWith({
-        test: {
-          query: "test",
-        },
+        test: {},
         test2: {},
       });
     });
