+++ public/pages/IndexDetail/containers/IndexDetail/index.ts
@@ -1,0 +1,3 @@
+import IndexDetail from "./IndexDetail";
+
+export default IndexDetail;
