+++ public/pages/IndexDetail/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import IndexDetail from "./containers/IndexDetail";
+
+export default IndexDetail;
