+++ public/pages/Indices/containers/IndexDetail/IndexDetail.test.tsx
@@ -6,7 +6,6 @@
 import React from "react";
 import "@testing-library/jest-dom/extend-expect";
 import { render, waitFor } from "@testing-library/react";
-import userEvent from "@testing-library/user-event";
 import { Route, Switch, HashRouter as Router } from "react-router-dom";
 import { browserServicesMock, coreServicesMock, apiCallerMock } from "../../../../../test/mocks";
 import IndexDetail, { IndexDetailModalProps } from "./index";
@@ -16,7 +15,7 @@ import { CoreServicesContext } from "../../../../components/core_services";
 
 apiCallerMock(browserServicesMock);
 
-function renderWithRouter(props: IndexDetailModalProps) {
+function renderWithRouter(props: Omit<IndexDetailModalProps, "history">) {
   return {
     ...render(
       <Router>
@@ -24,7 +23,7 @@ function renderWithRouter(props: IndexDetailModalProps) {
           <ServicesContext.Provider value={browserServicesMock}>
             <ModalProvider>
               <Switch>
-                <Route path="/" render={() => <IndexDetail {...props} />} />
+                <Route path="/" render={(routeProps) => <IndexDetail {...props} history={routeProps.history} />} />
               </Switch>
             </ModalProvider>
           </ServicesContext.Provider>
@@ -37,7 +36,7 @@ function renderWithRouter(props: IndexDetailModalProps) {
 describe("container <IndexDetail /> spec", () => {
   const onUpdateSuccessMock = jest.fn();
   it("render the component", async () => {
-    const { container, getByTestId, queryByText } = renderWithRouter({
+    const { container } = renderWithRouter({
       index: "test_index",
       record: {
         "docs.count": "5",
@@ -54,39 +53,11 @@ describe("container <IndexDetail /> spec", () => {
         managedPolicy: "",
         data_stream: "",
       },
-      onDelete: () => null,
       onUpdateIndex: onUpdateSuccessMock,
-      onClose: () => null,
-      onOpen: () => null,
-      onShrink: () => null,
     });
 
     await waitFor(() => {
       expect(container.firstChild).toMatchSnapshot();
-    });
-
-    userEvent.click(getByTestId("viewIndexDetailButton-test_index"));
-
-    await waitFor(() => {
-      expect(document.querySelector("#index-detail-modal-overview")).not.toBeNull();
-      expect(browserServicesMock.commonService.apiCaller).toBeCalledTimes(1);
-      expect(browserServicesMock.commonService.apiCaller).toBeCalledWith({
-        endpoint: "indices.get",
-        data: {
-          index: "test_index",
-        },
-      });
-    });
-
-    userEvent.click(document.getElementById("index-detail-modal-alias") as Element);
-    await waitFor(() => {
-      expect(queryByText("Index alias")).not.toBeNull();
-    });
-    userEvent.click(getByTestId("detail-modal-edit"));
-    await waitFor(() => {});
-    userEvent.click(getByTestId("createIndexCreateButton"));
-    await waitFor(() => {
-      expect(onUpdateSuccessMock).toBeCalledTimes(1);
     });
   });
 });
