+++ public/pages/Main/Main.tsx
@@ -38,6 +38,7 @@ import Reindex from "../Reindex/container/Reindex";
 import Aliases from "../Aliases";
 import Templates from "../Templates";
 import CreateIndexTemplate from "../CreateIndexTemplate";
+import IndexDetail from "../IndexDetail";
 
 enum Navigation {
   IndexManagement = "Index Management",
@@ -84,7 +85,7 @@ const HIDDEN_NAV_ROUTES = [
   ROUTES.CREATE_TEMPLATE,
 ];
 
-const HIDDEN_NAV_STARTS_WITH_ROUTE = [ROUTES.CREATE_TEMPLATE];
+const HIDDEN_NAV_STARTS_WITH_ROUTE = [ROUTES.CREATE_TEMPLATE, ROUTES.INDEX_DETAIL];
 
 interface MainProps extends RouteComponentProps {
   landingPage: string;
@@ -466,6 +467,14 @@ export default class Main extends Component<MainProps, object> {
                             render={(props) => (
                               <div style={ROUTE_STYLE}>
                                 <CreateIndexTemplate {...props} />
+                              </div>
+                            )}
+                          />
+                          <Route
+                            path={`${ROUTES.INDEX_DETAIL}/:index`}
+                            render={(props) => (
+                              <div style={ROUTE_STYLE}>
+                                <IndexDetail {...props} />
                               </div>
                             )}
                           />
