+++ public/utils/constants.ts
@@ -55,6 +55,7 @@ export const ROUTES = Object.freeze({
   CREATE_REPOSITORY: "/create-repository",
   EDIT_REPOSITORY: "/edit-repository",
   CREATE_INDEX: "/create-index",
+  INDEX_DETAIL: "/index-detail",
   REINDEX: "/reindex",
   ALIASES: "/aliases",
   TEMPLATES: "/templates",
@@ -95,6 +96,7 @@ export const BREADCRUMBS = Object.freeze({
   EDIT_REPOSITORY: { text: "Edit repository", href: `#${ROUTES.EDIT_REPOSITORY}` },
   CREATE_INDEX: { text: "Create Index", href: `#${ROUTES.CREATE_INDEX}` },
   EDIT_INDEX: { text: "Edit Index", href: `#${ROUTES.CREATE_INDEX}` },
+  INDEX_DETAIL: { text: "Index Detail", href: "#" },
   REINDEX: { text: "Reindex", href: `#${ROUTES.REINDEX}` },
   ALIASES: { text: "Aliases", href: `#${ROUTES.ALIASES}` },
   TEMPLATES: { text: "Templates", href: `#${ROUTES.TEMPLATES}` },
